refactor screenshot support names to be more generic to help out with future tests
