+++ WordPress/src/androidTest/java/org/wordpress/android/support/SupplierIdler.java
@@ -2,7 +2,7 @@ package org.wordpress.android.support;
 
 import java.util.function.Supplier;
 
-public class SupplierIdler extends WPScreenshotIdler {
+public class SupplierIdler extends WPIdler {
     private Supplier<Boolean> mSupplier;
 
     public SupplierIdler(Supplier<Boolean> supplier) {
