+++ WordPress/src/androidTest/java/org/wordpress/android/support/WPIdler.java
@@ -3,7 +3,7 @@ package org.wordpress.android.support;
 import android.support.test.espresso.IdlingResource;
 import static junit.framework.Assert.fail;
 
-public abstract class WPScreenshotIdler implements IdlingResource {
+public abstract class WPIdler implements IdlingResource {
     protected ResourceCallback mResourceCallback;
     protected Boolean mConditionWasMet = false;
