+++ WordPress/src/androidTest/java/org/wordpress/android/ui/screenshots/WPScreenshotTest.java
@@ -13,6 +13,7 @@ import org.wordpress.android.R;
 import org.wordpress.android.datasets.ReaderTagTable;
 import org.wordpress.android.models.ReaderTag;
 import org.wordpress.android.models.ReaderTagList;
+import org.wordpress.android.support.BaseTest;
 import org.wordpress.android.ui.WPLaunchActivity;
 import org.wordpress.android.ui.posts.EditPostActivity;
 import org.wordpress.android.ui.reader.views.ReaderSiteHeaderView;
@@ -23,30 +24,27 @@ import java.util.function.Supplier;
 import tools.fastlane.screengrab.Screengrab;
 import tools.fastlane.screengrab.UiAutomatorScreenshotStrategy;
 
-import static org.wordpress.android.BuildConfig.SCREENSHOT_LOGINPASSWORD;
-import static org.wordpress.android.BuildConfig.SCREENSHOT_LOGINUSERNAME;
-import static org.wordpress.android.support.WPScreenshotSupport.clickOn;
-import static org.wordpress.android.support.WPScreenshotSupport.clickOnChildAtIndex;
-import static org.wordpress.android.support.WPScreenshotSupport.focusEditPostTitle;
-import static org.wordpress.android.support.WPScreenshotSupport.getCurrentActivity;
-import static org.wordpress.android.support.WPScreenshotSupport.isElementDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.populateTextField;
-import static org.wordpress.android.support.WPScreenshotSupport.pressBackUntilElementIsDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.scrollToThenClickOn;
-import static org.wordpress.android.support.WPScreenshotSupport.selectItemAtIndexInSpinner;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForAtLeastOneElementOfTypeToBeDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForAtLeastOneElementWithIdToBeDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForConditionToBeTrue;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForElementToBeDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForElementToBeDisplayedWithoutFailure;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForElementToNotBeDisplayed;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForImagesOfTypeWithPlaceholder;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForRecyclerViewToStopReloading;
-import static org.wordpress.android.support.WPScreenshotSupport.waitForSwipeRefreshLayoutToStopReloading;
+import static org.wordpress.android.support.WPSupportUtils.clickOn;
+import static org.wordpress.android.support.WPSupportUtils.clickOnChildAtIndex;
+import static org.wordpress.android.support.WPSupportUtils.focusEditPostTitle;
+import static org.wordpress.android.support.WPSupportUtils.getCurrentActivity;
+import static org.wordpress.android.support.WPSupportUtils.isElementDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.pressBackUntilElementIsDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.scrollToThenClickOn;
+import static org.wordpress.android.support.WPSupportUtils.selectItemAtIndexInSpinner;
+import static org.wordpress.android.support.WPSupportUtils.waitForAtLeastOneElementOfTypeToBeDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.waitForAtLeastOneElementWithIdToBeDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.waitForConditionToBeTrue;
+import static org.wordpress.android.support.WPSupportUtils.waitForElementToBeDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.waitForElementToBeDisplayedWithoutFailure;
+import static org.wordpress.android.support.WPSupportUtils.waitForElementToNotBeDisplayed;
+import static org.wordpress.android.support.WPSupportUtils.waitForImagesOfTypeWithPlaceholder;
+import static org.wordpress.android.support.WPSupportUtils.waitForRecyclerViewToStopReloading;
+import static org.wordpress.android.support.WPSupportUtils.waitForSwipeRefreshLayoutToStopReloading;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class WPScreenshotTest {
+public class WPScreenshotTest extends BaseTest  {
     @ClassRule
     public static final WPLocaleTestRule LOCALE_TEST_RULE = new WPLocaleTestRule();
 
@@ -60,50 +58,13 @@ public class WPScreenshotTest {
         mActivityTestRule.launchActivity(null);
         Screengrab.setDefaultScreenshotStrategy(new UiAutomatorScreenshotStrategy());
 
-        wPLogin();
+        wpLogin();
+        sleep();
         editBlogPost();
         navigateReader();
         navigateNotifications();
         navigateStats();
-        wPLogout();
-    }
-
-    private void wPLogin() {
-        // If we're already logged in, log out before starting
-        if (!isElementDisplayed(R.id.login_button)) {
-            wPLogout();
-        }
-
-        // Login Prologue – We want to log in, not sign up
-        // See LoginPrologueFragment
-        clickOn(R.id.login_button);
-
-        // Email Address Screen – Fill it in and click "Next"
-        // See LoginEmailFragment
-        populateTextField(R.id.input, SCREENSHOT_LOGINUSERNAME);
-        clickOn(R.id.primary_button);
-
-        // Receive Magic Link or Enter Password Screen – Choose "Enter Password"
-        // See LoginMagicLinkRequestFragment
-        clickOn(R.id.login_enter_password);
-
-        // Password Screen – Fill it in and click "Next"
-        // See LoginEmailPasswordFragment
-        populateTextField(R.id.input, SCREENSHOT_LOGINPASSWORD);
-        clickOn(R.id.primary_button);
-
-        // Login Confirmation Screen – Click "Continue"
-        // See LoginEpilogueFragment
-        clickOn(R.id.primary_button);
-    }
-
-    private void wPLogout() {
-        // Click on the "Me" tab in the nav, then choose "Log Out"
-        clickOn(R.id.nav_me);
-        scrollToThenClickOn(R.id.row_logout);
-
-        // Confirm that we want to log out
-        clickOn(android.R.id.button1);
+        wpLogout();
     }
 
     private void navigateReader() {
